fix: create dirs if needed
checkForData() was simply looking for dirs to exist but we never created them. So it always failed.
Also it was silent because we were checking for the parent directory to exist (which it didn't) and
there was not logging on missing dirs. Now we ensure the dirs at the start of the function and then
do our checks.