feat(data-extraction): extract data from server.jar
Partially addresses #11. Utilizes the JarDataExtraction class to extract the stored data files in
the Minecraft server.jar file: advancements, loot_tables, recipes, structures, and tags. This
extraction can take a while, there are nearly 4000 files that make up this data. Only on the first
run will it have to extract these, subsequent runs it checks in the cache (work directory) for the
file structure and will skip the extraction step. We utilize node-7zip with bundled binary 7zip-bin
to handle the extraction. While the extraction is on-going, a log message gets printed every 5s. I
think eventually we will need to refactor so that extraction, level.dat conversion, and profile
grabbing happens before we begin processing other data (so that they can all factor in).

address #11