refactor: promises improvement
Perform promises in the main process. Also, if data extraction needs to happen it will do the
extraction then exit. You have to run mcdata-to-json again after the extraction completes to have it
run.

BREAKING CHANGE: If no extracted data, it will extract data and exit, requiring you to run it again.